Moved code to studentForm component and changed input style
